requirements.txt
Flask==2.3.3
Flask-Cors==3.1.2
gitpython==3.1.36
gunicorn==21.2.0
pytest==7.4.0
